backend/requirements.txt
aiopg==1.3.5
alembic==1.8.1
anyio==3.6.1
async-timeout==4.0.2
asyncpg==0.26.0
bcrypt==4.0.0
cffi==1.15.1
click==8.1.3
colorama==0.4.5
cryptography==38.0.1
databases==0.6.1
ecdsa==0.18.0
fastapi==0.85.0
greenlet==1.1.3
h11==0.14.0
httptools==0.5.0
idna==3.4
Mako==1.2.3
MarkupSafe==2.1.1
passlib==1.7.4
psycopg2-binary==2.9.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.20.4
typing_extensions==4.4.0
uvicorn==0.18.3
watchfiles==0.17.0
websockets==10.3
